backend/requirements.txt
Django
psycopg2-binary
djangorestframework==3.8.2
django-rest-auth
django-allauth
django-cors-headers
djangorestframework-gis==0.13
django-rest-swagger